+++ apps/api/src/factories/controllers/events-list-controller-factory.ts
@@ -1,5 +1,5 @@
-import { EventsListController } from '../../modules/events/EventsListController';
-import { EventsRepository } from '../../modules/events/repository/EventsRepository';
+import { EventsListController } from '../../modules/events/events-list-controller';
+import { EventsRepository } from '../../modules/events/repository/events-repository';
 import { EventsListService } from '../../modules/events/services';
 
 export const makeEventsListController = (): EventsListController => {
