+++ apps/api/src/modules/events/services/index.ts
@@ -1,1 +1,1 @@
-export * from './EventsListService';
+export * from './events-list-service';
